+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Output the mode for the functions `euidaccess()` and `faccessat()`
+- Output the three-dots `...` for the functions `futimesat()`, `lutimes()`,
+  `utime()`, `utimensat()`, and `utimes()`, 
 
 ### Fixed